pyproject.toml
[project]
name = "weather-forecast"
version = "0.5.0"
description = "Weather Forecast"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "dbt-duckdb>=1.10.0",
    "fastapi>=0.128.0",
    "pandas>=2.3.3",
    "plotly>=6.5.2",
    "polars>=1.37.1",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "streamlit>=1.53.0",
    "uvicorn>=0.40.0",
    "xgboost>=3.1.3",
]
